Update for Excella
Hi, Excella! This update is for you to see. It includes the latest
version of my code.

+++ src/main/java/me/happyman/utils/FTPAccessor.java
@@ -1,28 +1,29 @@
 package me.happyman.utils;
 
-import me.happyman.source;
+import me.happyman.worlds.UUIDFetcher;
 import org.apache.commons.net.ftp.FTPClient;
 import org.apache.commons.net.ftp.FTPFile;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
 
+import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
+
+import static me.happyman.Plugin.*;
 
 public class FTPAccessor
 {
     private static final int RECONNECT_DELAY = 200;
     private static final String CREDENTIAL_FILENAME = "credentials.txt";
-    private static source plugin;
     private static boolean isWorking = false;
     private static final FTPClient client = new FTPClient();
     private static Integer connectionAttemptTask = null;
 
-    public FTPAccessor(source plugin)
+    public FTPAccessor()
     {
         connectionAttemptTask = -1;
-        FTPAccessor.plugin = plugin;
         establishConnection();
         connectionAttemptTask = null;
     }
@@ -33,13 +34,15 @@ public class FTPAccessor
         {
             try
             {
-                if (!plugin.hasFile(DirectoryType.SERVER_DATA, "", CREDENTIAL_FILENAME))
+                File credentialFile = FileManager.getServerDataFile("", CREDENTIAL_FILENAME, false);
+                if (!credentialFile.exists())
                 {
-                    throw new IOException("You must put into your plugins/Smash/ServerData a "  + CREDENTIAL_FILENAME + " containing your FTP website login credentials like this:\n\tHostname: <hostname>\n\tUsername: <username>\n\tPassword: <password>");
+                    throw new IOException("You must put into your plugins/Smash/ServerData a "  + CREDENTIAL_FILENAME +
+                            " containing your FTP website login credentials like this:\n\tHostname: <hostname>\n\tUsername: <username>\n\tPassword: <password>");
                 }
-                String hostname = plugin.getDatumSimple(DirectoryType.SERVER_DATA, "", CREDENTIAL_FILENAME, "Hostname");
-                String username = plugin.getDatumSimple(DirectoryType.SERVER_DATA, "", CREDENTIAL_FILENAME, "Username");
-                String password = plugin.getDatumSimple(DirectoryType.SERVER_DATA, "", CREDENTIAL_FILENAME, "Password");
+                String hostname = FileManager.getSimpleData(credentialFile, "Hostname");
+                String username = FileManager.getSimpleData(credentialFile, "Username");
+                String password = FileManager.getSimpleData(credentialFile, "Password");
                 if (hostname.length() > 1 && username.equals("") || username.equals("*") || password.equals("") || password.equals("*"))
                 {
                     throw new IOException("Credentials file not filled out!");
@@ -54,12 +57,12 @@ public class FTPAccessor
                 {
                     throw new IOException("Please verify login credentials!");
                 }
-                Bukkit.getConsoleSender().sendMessage(plugin.loggerPrefix() + ChatColor.GREEN + "FTP connection established!");
+                Bukkit.getConsoleSender().sendMessage(loggerPrefix() + ChatColor.GREEN + "FTP connection established!");
                 isWorking = true;
             }
             catch (IOException ex)
             {
-                plugin.sendErrorMessage("Error! FTP connection failed! " + ex.getMessage());
+                sendErrorMessage("Error! FTP connection failed! " + ex.getMessage());
                 isWorking = false;
             }
         }
@@ -78,7 +81,7 @@ public class FTPAccessor
     {
         if (!isWorking && connectionAttemptTask == null)
         {
-            connectionAttemptTask = Bukkit.getScheduler().scheduleSyncDelayedTask(plugin, new Runnable()
+            connectionAttemptTask = Bukkit.getScheduler().scheduleSyncDelayedTask(getPlugin(), new Runnable()
             {
                 public void run()
                 {
@@ -89,33 +92,37 @@ public class FTPAccessor
         }
     }
 
-    public static void copyTextFileIntoFTP(DirectoryType sourceDirectory, String relativePath, String sourceFileName, String relPathInFTP)
-    {
+    public static void copyTextFileIntoFTP(File fileToCopy, String relPathInFTP)
+    {//
         if (!isWorking)
         {
-            plugin.sendErrorMessage("Error! FTP connection has not been established!");
+            sendErrorMessage("Error! FTP connection has not been established!");
         }
         else
         {
-            if (!sourceFileName.contains(".") && sourceFileName.length() > 0)
+            boolean bad = false;
+            if (fileToCopy != null && fileToCopy.exists())
             {
-                sourceFileName += ".json";
-            }
+                String absoluteDestinationDirectory = FileManager.desandwichPathInSlashes(relPathInFTP) + '/' + fileToCopy.getName();
 
-            String absoluteSourceDirectory = plugin.getAbsolutePath(sourceDirectory, relativePath) + sourceFileName;
-            String absoluteDestinationDirectory = source.makeForwardAndSandwichInSlashes(relPathInFTP) + sourceFileName;
-
-            try
-            {
-                if (!client.storeFile(absoluteDestinationDirectory, new FileInputStream(absoluteSourceDirectory)))
+                try
                 {
-                    plugin.sendErrorMessage("Unable to store the file " + sourceFileName + " to the FTP! Be sure that the directory "
-                    + ChatColor.YELLOW + absoluteDestinationDirectory.substring(0, absoluteDestinationDirectory.lastIndexOf('/')) + ChatColor.RED + " exists.");
+                    if (!client.storeFile(absoluteDestinationDirectory, new FileInputStream(absoluteDestinationDirectory)))
+                    {
+                        sendErrorMessage("Unable to store the file " + fileToCopy.getName() + " to the FTP! Be sure that the directory " +
+                                ChatColor.YELLOW + relPathInFTP + ChatColor.RED + " exists over there.");
+                    }
                 }
+                catch (IOException ex)
+                {
+                    ex.printStackTrace();
+                    bad = true;
+                }
+
             }
-            catch (IOException ex)
+            if (bad)
             {
-                plugin.sendErrorMessage("Could not find the file " + sourceFileName + " when trying to send to the website!");
+                sendErrorMessage("Could not find the file when trying to send to the website!");
             }
         }
     }
@@ -129,31 +136,34 @@ public class FTPAccessor
     {
         if (!isWorking)
         {
-            plugin.sendErrorMessage("Error! Attempted to save a player profile while FTP was not established!");
+            sendErrorMessage("Error! Attempted to saveData a player profile while FTP was not established!");
             establishConnection();
         }
         else
         {
             String relPath = "/public_html/documents/player_data";
-            String uuid = plugin.getUUID(p);
-            try
+            String uuid = UUIDFetcher.getUUID(p);
+            if (uuid != null)
             {
-                FTPFile[] fileList = client.listFiles("/" + relPath);
-                for (int i = 0; i < fileList.length; i++)
+                try
                 {
-                    if (fileList[i].getName().contains(uuid))
+                    FTPFile[] fileList = client.listFiles("/" + relPath);
+                    for (int i = 0; i < fileList.length; i++)
                     {
-                        client.deleteFile("/" + relPath + "/" + fileList[i].getName());
-                        break;
+                        if (fileList[i].getName().contains(uuid))
+                        {
+                            client.deleteFile("/" + relPath + "/" + fileList[i].getName());
+                            break;
+                        }
                     }
                 }
+                catch (IOException ex)
+                {
+                    sendErrorMessage(ChatColor.RED + "Error! FTP " + ex.getMessage().toLowerCase() + "!");
+                    isWorking = false;
+                }
+                copyTextFileIntoFTP(FileManager.getGeneralPlayerFile(p), relPath);
             }
-            catch (IOException ex)
-            {
-                plugin.sendErrorMessage(ChatColor.RED + "Error! FTP " + ex.getMessage().toLowerCase() + "!");
-                isWorking = false;
-            }
-            copyTextFileIntoFTP(DirectoryType.PLAYER_DATA, "", plugin.getPlayerFileName(p), relPath);
         }
     }
 }
